+++ adapters/IronSource/CHANGELOG.md
@@ -1,5 +1,9 @@
 # ironSource Adapter for Google Mobile Ads SDK for iOS
 
-## 6.7.5.0
+
+## 2.6.1
+- Initiate IronSource SDK once per ad unit
+
+## 2.6.0
 - Initial release!
 - Supports interstitial and rewarded video ads.